added rudimentary hero moving through new widget RTSSnapToGridBP
Utilising in built snap to grid functionality I've added hero moving through the widget RTSSnapToGridBP.

Side note: I added a new boolean that exists in SnapToGrid to tell RTSPlayerControllerBP that it is currently hitting an object and to stop moving the CameraPawn.

Issues: As soon as the object snaps to a grid far enough away it is activating the camerapawn again which is not ideal. Need to figure out better functionality for that.