functionality added to rtssnaptogrid and changed triggerboxes to a seperate BP
spawning heroes onto the sideboard now works as intended

need to make functions to do checks for if the sideboard is full to determine if a player can make a purchase.

next step is to make that function and provide users a widget that allows them to buy from a randomised hero pool every round. (gotta check that the sideboard is not full before taking money away)